Refator: 코드가 많지 않고 연관성 있는 모델이라 통합
